+++ docs/CHANGELOG.md
@@ -10,6 +10,10 @@ title: Changelog
 * Set maximum line length to 120.
 
     *Joel Hawksley*
+    
+* Setting a collection slot with the plural setter (`component.items(array)` for `renders_many :items`)  returns the array of slots.
+
+    *Jon Palmer*
 
 * Update error messages to be more descriptive and helpful.